+++ build_modules/test/modules_test.dart
@@ -2,11 +2,16 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import 'dart:convert';
+import 'dart:io';
+
 import 'package:analyzer/dart/element/element.dart';
 import 'package:build_test/build_test.dart';
 import 'package:test/test.dart';
 
+import 'package:build_modules/src/errors.dart';
 import 'package:build_modules/src/modules.dart';
+import 'package:build_modules/src/module_builder.dart';
 
 void main() {
   LibraryElement libCycle;
@@ -15,6 +20,8 @@ void main() {
   LibraryElement libCycleWithB;
   LibraryElement libCycleWithA;
   LibraryElement libDepOnNonSdk;
+  LibraryElement libAImportsBNoCycle;
+  LibraryElement libBImportsANoCycle;
   final assetCycle = makeAssetId('a|lib/a_cycle.dart');
   final assetSecondaryInCycle = makeAssetId('a|lib/a_secondary_in_cycle.dart');
   final assetPartInCycle = makeAssetId('a|lib/a_part_in_cycle.dart');
@@ -23,6 +30,8 @@ void main() {
   final assetCycleWithA = makeAssetId('b|lib/b_cycle_with_a.dart');
   final assetDepOnNonSdk = makeAssetId('a|lib/a_dep_on_non_sdk.dart');
   final assetNonSdk = makeAssetId('a|lib/a_non_sdk.dart');
+  final assetAImportsBNoCycle = makeAssetId('a|lib/a_imports_b_no_cycle.dart');
+  final assetBImportsANoCycle = makeAssetId('b|lib/b_imports_a_no_cycle.dart');
 
   setUpAll(() async {
     await resolveAsset(assetCycle, (resolver) async {
@@ -38,6 +47,12 @@ void main() {
     });
     await resolveAsset(assetDepOnNonSdk, (resolver) async {
       libDepOnNonSdk = await resolver.libraryFor(assetDepOnNonSdk);
+    });
+    await resolveAsset(assetAImportsBNoCycle, (resolver) async {
+      libAImportsBNoCycle = await resolver.libraryFor(assetAImportsBNoCycle);
+    });
+    await resolveAsset(assetBImportsANoCycle, (resolver) async {
+      libBImportsANoCycle = await resolver.libraryFor(assetBImportsANoCycle);
     });
   });
 
@@ -97,6 +112,62 @@ void main() {
       test('Does not mark library with later alpha sort as primary', () {
         expect(isPrimary(libCycleWithA), false);
       });
+    });
+  });
+
+  group('computeTransitiveDeps', () {
+    Module rootModule;
+    Module immediateDep;
+    Module transitiveDep;
+    InMemoryAssetReader reader;
+
+    setUp(() {
+      rootModule = new Module.forLibrary(libAImportsBNoCycle);
+      immediateDep = new Module.forLibrary(libBImportsANoCycle);
+      transitiveDep = new Module.forLibrary(libNoCycle);
+      reader = new InMemoryAssetReader();
+      reader.cacheStringAsset(
+          assetAImportsBNoCycle,
+          new File('test/fixtures/a/lib/a_imports_b_no_cycle.dart')
+              .readAsStringSync());
+      reader.cacheStringAsset(
+          assetBImportsANoCycle,
+          new File('test/fixtures/b/lib/b_imports_a_no_cycle.dart')
+              .readAsStringSync());
+    });
+
+    test('finds transitive deps', () async {
+      reader.cacheStringAsset(
+          assetBImportsANoCycle.changeExtension(moduleExtension),
+          JSON.encode(immediateDep.toJson()));
+      reader.cacheStringAsset(assetNoCycle.changeExtension(moduleExtension),
+          JSON.encode(transitiveDep.toJson()));
+
+      var transitiveDeps =
+          await rootModule.computeTransitiveDependencies(reader);
+      expect(
+          transitiveDeps.map((m) => m.primarySource),
+          unorderedEquals(
+              [immediateDep.primarySource, transitiveDep.primarySource]));
+    });
+
+    test('missing modules report nice errors', () {
+      reader.cacheStringAsset(
+          assetBImportsANoCycle.changeExtension(moduleExtension),
+          JSON.encode(immediateDep.toJson()));
+      expect(
+          () => rootModule.computeTransitiveDependencies(reader),
+          allOf(throwsA(new isInstanceOf<MissingModulesException>()), throwsA(
+            predicate<MissingModulesException>(
+              (error) {
+                return error.message.contains('''
+Unable to find modules for some sources, check the following imports:
+
+`import 'package:a/a_no_cycle.dart';` from b|lib/b_imports_a_no_cycle.dart at 2:1
+''');
+              },
+            ),
+          )));
     });
   });
 }