+++ build_web_compilers/test/dev_compiler_builder_test.dart
@@ -3,6 +3,7 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:build_test/build_test.dart';
+import 'package:logging/logging.dart';
 import 'package:test/test.dart';
 
 import 'package:build_web_compilers/build_web_compilers.dart';
@@ -13,39 +14,102 @@ import 'util.dart';
 main() {
   Map<String, dynamic> assets;
 
-  setUp(() async {
-    assets = {
-      'build_modules|lib/src/analysis_options.default.yaml': '',
-      'b|lib/b.dart': '''final world = 'world';''',
-      'a|lib/a.dart': '''
+  group('error free project', () {
+    setUp(() async {
+      assets = {
+        'build_modules|lib/src/analysis_options.default.yaml': '',
+        'b|lib/b.dart': '''final world = 'world';''',
+        'a|lib/a.dart': '''
         import 'package:b/b.dart';
         final hello = world;
       ''',
-      'a|web/index.dart': '''
+        'a|web/index.dart': '''
         import "package:a/a.dart";
         main() {
           print(hello);
         }
       ''',
-    };
+      };
 
-    // Set up all the other required inputs for this test.
-    await testBuilderAndCollectAssets(new ModuleBuilder(), assets);
-    await testBuilderAndCollectAssets(new UnlinkedSummaryBuilder(), assets);
-    await testBuilderAndCollectAssets(new LinkedSummaryBuilder(), assets);
+      // Set up all the other required inputs for this test.
+      await testBuilderAndCollectAssets(new ModuleBuilder(), assets);
+      await testBuilderAndCollectAssets(new UnlinkedSummaryBuilder(), assets);
+      await testBuilderAndCollectAssets(new LinkedSummaryBuilder(), assets);
+    });
+
+    test('can compile ddc modules under lib and web', () async {
+      var expectedOutputs = {
+        'b|lib/b$jsModuleExtension': decodedMatches(contains('world')),
+        'b|lib/b$jsSourceMapExtension': decodedMatches(contains('b.dart')),
+        'a|lib/a$jsModuleExtension': decodedMatches(contains('hello')),
+        'a|lib/a$jsSourceMapExtension': decodedMatches(contains('a.dart')),
+        'a|web/index$jsModuleExtension': decodedMatches(contains('main')),
+        'a|web/index$jsSourceMapExtension':
+            decodedMatches(contains('index.dart')),
+      };
+      await testBuilder(new DevCompilerBuilder(), assets,
+          outputs: expectedOutputs);
+    });
   });
 
-  test("can compile ddc modules under lib and web", () async {
-    var expectedOutputs = {
-      'b|lib/b$jsModuleExtension': decodedMatches(contains('world')),
-      'b|lib/b$jsSourceMapExtension': decodedMatches(contains('b.dart')),
-      'a|lib/a$jsModuleExtension': decodedMatches(contains('hello')),
-      'a|lib/a$jsSourceMapExtension': decodedMatches(contains('a.dart')),
-      'a|web/index$jsModuleExtension': decodedMatches(contains('main')),
-      'a|web/index$jsSourceMapExtension':
-          decodedMatches(contains('index.dart')),
-    };
-    await testBuilder(new DevCompilerBuilder(), assets,
-        outputs: expectedOutputs);
+  group('projects with errors due to', () {
+    group('invalid assignements', () {
+      setUp(() async {
+        assets = {
+          'build_modules|lib/src/analysis_options.default.yaml': '',
+          'a|web/index.dart': 'int x = "hello";',
+        };
+
+        // Set up all the other required inputs for this test.
+        await testBuilderAndCollectAssets(new ModuleBuilder(), assets);
+        await testBuilderAndCollectAssets(new UnlinkedSummaryBuilder(), assets);
+        await testBuilderAndCollectAssets(new LinkedSummaryBuilder(), assets);
+      });
+
+      test('reports useful messages', () async {
+        var expectedOutputs = {
+          'a|web/index$jsModuleErrorsExtension': decodedMatches(
+              allOf(contains('String'), contains('assigned'), contains('int'))),
+        };
+        var logs = <LogRecord>[];
+        await testBuilder(new DevCompilerBuilder(), assets,
+            outputs: expectedOutputs, onLog: logs.add);
+        expect(
+            logs,
+            contains(predicate<LogRecord>((record) =>
+                record.level == Level.SEVERE &&
+                record.message.contains('String') &&
+                record.message.contains('assigned') &&
+                record.message.contains('int'))));
+      });
+    });
+
+    group('invalid imports', () {
+      setUp(() async {
+        assets = {
+          'build_modules|lib/src/analysis_options.default.yaml': '',
+          'a|web/index.dart': "import 'package:a/a.dart'",
+        };
+
+        // Set up all the other required inputs for this test.
+        await testBuilderAndCollectAssets(new ModuleBuilder(), assets);
+      });
+
+      test('reports useful messages', () async {
+        var expectedOutputs = {
+          'a|web/index$jsModuleErrorsExtension': decodedMatches(
+              contains('Unable to find modules for some sources')),
+        };
+        var logs = <LogRecord>[];
+        await testBuilder(new DevCompilerBuilder(), assets,
+            outputs: expectedOutputs, onLog: logs.add);
+        expect(
+            logs,
+            contains(predicate<LogRecord>((record) =>
+                record.level == Level.SEVERE &&
+                record.message
+                    .contains('Unable to find modules for some sources'))));
+      });
+    });
   });
 }